requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==3.14.0
psycopg2-binary==2.9.1
